Migrates to Avro serialization
Updates producer to use Avro serialization with Schema Registry.

This change improves data compatibility and schema evolution by leveraging Avro's schema management capabilities.
