+++ KafkaFlow/KafkaFlowConsumer/Handlers/TrainingMessageHandler.cs
