+++ KafkaFlow/KafkaFlowProducer/KafkaFlowProducer.csproj
@@ -11,14 +11,15 @@
         <PackageReference Include="Confluent.SchemaRegistry.Serdes.Avro" Version="2.8.0" />
         <PackageReference Include="KafkaFlow" Version="3.1.0" />
         <PackageReference Include="KafkaFlow.Microsoft.DependencyInjection" Version="3.1.0" />
-        <PackageReference Include="KafkaFlow.Serializer.JsonCore" Version="3.1.0" />
-        <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="9.0.0-rc.1.24452.1"/>
+        <PackageReference Include="KafkaFlow.Serializer.SchemaRegistry.ConfluentAvro" Version="4.0.0" />
+        <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="9.0.0-rc.1.24452.1" />
         <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="9.0.3">
           <PrivateAssets>all</PrivateAssets>
           <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
         <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="9.0.3" />
         <PackageReference Include="SQLite" Version="3.13.0" />
+        <PackageReference Include="Streamiz.Kafka.Net.SchemaRegistry.SerDes.Avro" Version="1.7.0" />
     </ItemGroup>
 
     <ItemGroup>
