+++ KafkaFlow/Models/Constants.cs
