+++ KafkaFlow/Models/TodoEvent.cs
