+++ KafkaFlow/Models/TrainingTodoEvent.cs
